Initial repositioning into feature folders with file renames
