--- src/components/app/login.css
@@ -1,25 +1,0 @@
-.loginContainer {
-  grid-row-start: 2;
-  grid-column-start: 4;
-  grid-column-end: 6;
-}
-
-.loginButtons {
-  height: 100%;
-}
-
-.loginButtons button {
-  border: 1px solid #56CCF2;
-  color: #56CCF2;
-  background-color: transparent;
-  height: 100%;
-  min-width: 90px;
-  margin-right: 0.5em;
-  font-family: 'roboto condensed';
-}
-
-.loginButtons button:hover {
-  background-color: #56CCF2;
-  color: #100E14;
-  cursor: pointer;
-}
