--- src/containers/simulator/simulatorStatus.css
@@ -1,63 +1,0 @@
-
-#simulatorStatus {
-
-  grid-row-start: 2;
-  grid-row-end: 3;
-  grid-column-start: 1;
-  grid-column-end: 3;
-
-  border: 2px solid #56CCF2;
-  border-bottom: none;
-}
-
-.resultContainer {
-  display: inline-block;
-  height: 23px;
-  font-family: 'roboto condensed';
-  color: #EB5757;
-  text-transform: uppercase;
-  margin-left: 10px;
-  margin-top: 5px;
-}
-
-@keyframes fadeInOut { /* flame pulses */
-  0%   { opacity:1; }
-  50%  { opacity:0; }
-  100% { opacity:1; }
-}
-
-.fade {
-    opacity: 1;
-    animation: fadeInOut 1s infinite;
-}
-
-.resultContainer div {
-  display: inline-block;
-  width: 100px;
-  margin: 0 10px 0 10px;
-  height: 16px;
-}
-
-div.winner_0 {
-  background-color: #EB5757;
-}
-
-div.winner_1 {
-  background-color: #6FCF97;
-}
-
-div.winner_2 {
-  background-color: #56CCF2;
-}
-
-div.winner_3 {
-  background-color: #F2C94C;
-}
-
-div.winner_4 {
-  background-color: #BB6BD9;
-}
-
-div.winner_5 {
-  background-color: #BDBDBD;
-}
