+++ src/features/simulator/core.js
