+++ src/features/simulator/fontAwesomeButton.js
