+++ src/features/simulator/runProgressIndicator.js
