+++ src/features/simulator/simulatorLayout.js
