+++ src/features/simulator/speedControl.js
