+++ src/features/simulator/warriors.js
