+++ src/reducers/index.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
 
-import parser from './parserReducers'
-import simulator from './simulatorReducers'
+import parser from '../features/parser/reducer'
+import simulator from '../features/simulator/reducer'
 
 export default combineReducers({
   routing: routerReducer,
