+++ src/sagas/index.js
@@ -1,7 +1,7 @@
 import { all } from 'redux-saga/effects'
 
-import { parserWatchers } from './parserSagas'
-import { simulatorWatchers } from './simulatorSagas'
+import { parserWatchers } from '../features/parser/sagas'
+import { simulatorWatchers } from '../features/simulator/sagas'
 
 export default function* rootSaga() {
   yield all([
